requirements.txt
faker==24.0.0
fastapi==0.110.0
uvicorn==0.27.0
confluent-kafka==2.3.0
duckdb==1.1.0
dbt-duckdb==1.8.0
# apache-airflow==2.8.2 # installed separately with constraint
great-expectations==0.18.12
streamlit==1.32.0
pandas==2.2.0
pyarrow==15.0.0
python-dotenv==1.0.1
httpx==0.27.0
